tutorials/google-adk/06_loop_agents/pyproject.toml
[project]
name = "06-loop-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
]
